refactor : H2 Db test
