+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'shopping-prcatice'
+rootProject.name = 'shopping-practice'
