+++ src/main/java/dongwook/shoppingpractice/domain/member/model/CurrentMember.java
@@ -1,0 +1,14 @@
+package dongwook.shoppingpractice.domain.member.model;
+
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME) // (1)
+@Target(ElementType.PARAMETER) // (2)
+@AuthenticationPrincipal(expression = "#this == 'anonymousUser' ? null : member")
+public @interface CurrentMember {
+}
