+++ src/main/java/dongwook/shoppingpractice/domain/member/model/UserMember.java
@@ -1,0 +1,18 @@
+package dongwook.shoppingpractice.domain.member.model;
+
+import lombok.Getter;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+
+import java.util.List;
+
+@Getter
+public class UserMember extends User {
+
+    private Member member;
+
+    public UserMember(Member member) {
+        super(member.getEmail(), member.getPassword(), List.of(new SimpleGrantedAuthority("USER")));
+        this.member = member;
+    }
+}
