--- src/main/java/dongwook/shoppingprcatice/ShoppingPrcaticeApplication.java
@@ -1,13 +1,0 @@
-package dongwook.shoppingprcatice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ShoppingPrcaticeApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ShoppingPrcaticeApplication.class, args);
-    }
-
-}
