--- src/main/java/dongwook/shoppingprcatice/domain/member/model/CurrentMember.java
@@ -1,14 +1,0 @@
-package dongwook.shoppingprcatice.domain.member.model;
-
-import org.springframework.security.core.annotation.AuthenticationPrincipal;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME) // (1)
-@Target(ElementType.PARAMETER) // (2)
-@AuthenticationPrincipal(expression = "#this == 'anonymousUser' ? null : member")
-public @interface CurrentMember {
-}
