--- src/main/java/dongwook/shoppingprcatice/domain/member/model/Member.java
@@ -1,31 +1,0 @@
-package dongwook.shoppingprcatice.domain.member.model;
-
-import dongwook.shoppingprcatice.domain.member.form.SignUpForm;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class Member {
-
-    private Long id;
-    private String username;
-    private String phoneNumber;
-    private String email;
-    private String zipcode;
-    private String password;
-    private String address;
-    private String addressDetail;
-
-
-    public Member(SignUpForm form) {
-        this.username = form.getUsername();
-        this.phoneNumber = form.getPhoneNumber();
-        this.email = form.getEmail();
-        this.zipcode = form.getZipcode();
-        this.password = form.getPassword();
-        this.address = form.getAddress();
-        this.addressDetail = form.getAddressDetail();
-    }
-
-}
