--- src/main/java/dongwook/shoppingprcatice/domain/member/model/UserMember.java
@@ -1,18 +1,0 @@
-package dongwook.shoppingprcatice.domain.member.model;
-
-import lombok.Getter;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-
-import java.util.List;
-
-@Getter
-public class UserMember extends User {
-
-    private Member member;
-
-    public UserMember(Member member) {
-        super(member.getEmail(), member.getPassword(), List.of(new SimpleGrantedAuthority("USER")));
-        this.member = member;
-    }
-}
