--- src/main/java/dongwook/shoppingprcatice/infra/config/security/SecurityUtils.java
@@ -1,15 +1,0 @@
-package dongwook.shoppingprcatice.infra.config.security;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.crypto.factory.PasswordEncoderFactories;
-import org.springframework.security.crypto.password.PasswordEncoder;
-
-@Configuration
-public class SecurityUtils {
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return PasswordEncoderFactories.createDelegatingPasswordEncoder();
-    }
-}
