+++ src/test/java/dongwook/shoppingpractice/ShoppingPracticeApplicationTests.java
@@ -1,0 +1,14 @@
+package dongwook.shoppingpractice;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ShoppingPracticeApplicationTests {
+
+    @Test
+    void contextLoads() {
+
+    }
+
+}
