--- src/test/java/dongwook/shoppingprcatice/ShoppingPrcaticeApplicationTests.java
@@ -1,13 +1,0 @@
-package dongwook.shoppingprcatice;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ShoppingPrcaticeApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
